issue-112 - continuing the module modularisation, fixing client unit tests

+++ client/src/test/java/com/github/goldy1992/mp3player/client/views/TimeCounterTest.kt
@@ -39,7 +39,7 @@ class TimeCounterTest {
         val expectedState = PlaybackStateCompat.STATE_PLAYING
         timeCounter!!.init(null)
         timeCounter!!.updateState(createState(expectedState, 5L))
-        Assert.assertEquals(expectedState.toLong(), timeCounter!!.currentState)
+        Assert.assertEquals(expectedState, timeCounter!!.currentState)
         // current position will not be updated will not updated as text view is null
         Assert.assertEquals(expectedPosition, timeCounter!!.currentPosition)
     }
